cleanup(5/n): relocate rupture_detail + apply_geojson_style graphene-free
The strawberry schema imported two pure (pandas/dict) helpers through graphene-
importing modules. Move them to graphene-free homes so the schema names no graphene
module:
- apply_geojson_style -> new geojson_style_util.py (geojson_style.py re-exports it)
- rupture_detail -> cached.py (composite_rupture_detail.py re-exports it)

strawberry_schema now imports both from the graphene-free modules. The legacy graphene
modules keep working via re-export shims until they are deleted at cutover (the package
__init__ still eager-loads them, which is removed then).

SDL byte-identical; 12/12 differential parity green.
